restructure file tree to be more specific
