+++ src/app/_components/icon/icon.component.html
