+++ src/app/_components/icon/icon.component.spec.ts
