+++ src/app/_components/icon/icon.component.ts
