+++ src/app/_components/logo/logo.component.scss
