+++ src/app/_components/logo/logo.component.spec.ts
