+++ src/app/_models/associate.ts
