+++ src/app/_models/spark-type.ts
