+++ src/app/_models/spark-types.ts
